docs: remove memory usage comment
this doesnt really use a large amount of memory
in the common case. If you dont supply a destination,
then it copies the db into memory, but thats what you
wanted. If you specify a destination, this uses
conn.backup, which is efficient -- it uses
the underlying C code and copies individual
database pages at a time